+++ packages/stryker-wct-runner/src/index.ts
@@ -1,0 +1,4 @@
+import { TestRunnerFactory } from 'stryker-api/test_runner';
+import WctTestRunner from './WctTestRunner';
+
+TestRunnerFactory.instance().register('wct', WctTestRunner);