+++ packages/stryker-wct-runner/test/helpers/sinonSetup.ts
@@ -1,0 +1,5 @@
+import * as sinon from 'sinon';
+
+afterEach(() => {
+  sinon.restore();
+});