+++ packages/stryker-jasmine-runner/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.9"></a>
+## [0.2.9](https://github.com/stryker-mutator/stryker/compare/stryker-jasmine-runner@0.2.8...stryker-jasmine-runner@0.2.9) (2018-12-12)
+
+
+
+
+**Note:** Version bump only for package stryker-jasmine-runner
+
 <a name="0.2.8"></a>
 ## [0.2.8](https://github.com/stryker-mutator/stryker/compare/stryker-jasmine-runner@0.2.7...stryker-jasmine-runner@0.2.8) (2018-11-29)