+++ packages/stryker-mocha-framework/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.13.3"></a>
+## [0.13.3](https://github.com/stryker-mutator/stryker/compare/stryker-mocha-framework@0.13.2...stryker-mocha-framework@0.13.3) (2018-12-12)
+
+
+
+
+**Note:** Version bump only for package stryker-mocha-framework
+
 <a name="0.13.2"></a>
 ## [0.13.2](https://github.com/stryker-mutator/stryker/compare/stryker-mocha-framework@0.13.1...stryker-mocha-framework@0.13.2) (2018-11-29)