+++ packages/stryker-mocha-runner/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.15.3"></a>
+## [0.15.3](https://github.com/stryker-mutator/stryker/compare/stryker-mocha-runner@0.15.2...stryker-mocha-runner@0.15.3) (2018-12-12)
+
+
+
+
+**Note:** Version bump only for package stryker-mocha-runner
+
 <a name="0.15.2"></a>
 ## [0.15.2](https://github.com/stryker-mutator/stryker/compare/stryker-mocha-runner@0.15.1...stryker-mocha-runner@0.15.2) (2018-11-29)