+++ packages/stryker-typescript/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.16.2"></a>
+## [0.16.2](https://github.com/stryker-mutator/stryker/compare/stryker-typescript@0.16.1...stryker-typescript@0.16.2) (2018-12-12)
+
+
+
+
+**Note:** Version bump only for package stryker-typescript
+
 <a name="0.16.1"></a>
 ## [0.16.1](https://github.com/stryker-mutator/stryker/compare/stryker-typescript@0.16.0...stryker-typescript@0.16.1) (2018-11-29)